fix: resolve open SonarCloud reliability issues
- S2789: replace the null-assigned Optional in
  LockAndBanListener.onVehicleMove with Optional.empty() plus an
  islandResolved flag, keeping the lazy single lookup per vehicle
- S2583: drop the @NonNull generic annotations on IslandCache#grids -
  Map#get legitimately returns null for worlds with no grid, and the
  value annotation made those null checks look impossible
- S2637: pass Objects.requireNonNull(p.getLocation()) to getIslandAt in
  InvincibleVisitorsListener - a live player's location is never null
- S2184: cast the chunk coordinate additions to double in
  ClosestSafeSpotTeleport#scanAndPopulateBlockQueue before the implicit
  widening in BoundingBox#contains

+++ src/main/java/world/bentobox/bentobox/util/teleport/ClosestSafeSpotTeleport.java
@@ -347,7 +347,7 @@ public class ClosestSafeSpotTeleport
                 {
                     // Raw-coordinate bounds check first and only allocate a vector for
                     // positions actually inside the search area
-                    if (this.boundingBox.contains(chunkX + x, y, chunkZ + z))
+                    if (this.boundingBox.contains((double) chunkX + x, y, (double) chunkZ + z))
                     {
                         // Process positions that are inside bounding box of search area.
                         Vector positionVector = new Vector(chunkX + x, y, chunkZ + z);
